+++ sw.js
@@ -2,7 +2,7 @@
 //
 // Bump CACHE version whenever shipped files change so clients pick up the update.
 
-const CACHE = 'tripletown-v29';
+const CACHE = 'tripletown-v30';
 
 const SHELL = [
   '.',
@@ -20,6 +20,7 @@ const SHELL = [
   'js/game.js',
   'js/render.js',
   'js/sprites.js',
+  'js/gestures.js',
   'assets/icons/icon-192.png',
   'assets/icons/icon-512.png',
   'assets/icons/icon-180.png',